Feature (Testing): Structured tests
Restructured testing folder.
