+++ testing/commands/ping.js
